Revert "sd-bus: make rqueue/wqueue sizes of type size_t"
This reverts commit 6ca588f3d62ad3391d558e34a7f002c0a88698b7.

+++ src/libsystemd/sd-bus/bus-internal.h
@@ -221,11 +221,11 @@ struct sd_bus {
         size_t rbuffer_size;
 
         sd_bus_message **rqueue;
-        size_t rqueue_size;
+        unsigned rqueue_size;
         size_t rqueue_allocated;
 
         sd_bus_message **wqueue;
-        size_t wqueue_size;
+        unsigned wqueue_size;
         size_t windex;
         size_t wqueue_allocated;
